+++ src/firebase/firebase.js
@@ -27,25 +27,7 @@ export const verifyEmail = () => {
   user.sendEmailVerification();
 }
 ///////////////////////////////////////////////////////////////////////////////////
-const getToken = async() => {
+export const getToken = async() => {
   const token = await auth.currentUser.getIdToken(true);
   return token
-}
-
-
-export const grantDefaultRole = async() => {
-
-  const token = await getToken();
-
-  let url = "http://192.168.1.13:5000/grantRole";
-
-  const headers = {
-    'Content-Type': 'application/json',
-    'Authorization': token
-  }
-
-  // Axios is used instead of fetch as it gave me weird results
-  let data = await axios.get(url, {headers: headers}).then(response => response.data);
-  console.log(data);
-  //return data;
 }
